moved common macros to common header
